This commit moves test_credentials.js to var/casper and updates its related reference.
tools/setup/generate-test-credentials: Move test_credentials to var/casper.

tools/setup/generate-test-credentials: Create directory var/casper.

frontend_tests/casper_lib/common.js: Use test_credentials from var/casper.

casper_tests/06-settings.js: Use test_credentials from var/casper.

casper_tests/10-admin.js: use test_credentials from var/casper.

gitignore: Remove frontend_tests/casper_lib/test_credentials.js .
