+++ frontend_tests/casper_lib/common.js
@@ -2,7 +2,7 @@ var common = (function () {
 
 var exports = {};
 
-var test_credentials = require('../casper_lib/test_credentials.js').test_credentials;
+var test_credentials = require('../../var/casper/test_credentials.js').test_credentials;
 
 function timestamp() {
     return new Date().getTime();
