+++ frontend_tests/casper_tests/10-admin.js
@@ -1,5 +1,5 @@
 var common = require('../casper_lib/common.js').common;
-var test_credentials = require('../casper_lib/test_credentials.js').test_credentials;
+var test_credentials = require('../../var/casper/test_credentials.js').test_credentials;
 
 common.start_and_log_in();
