+++ tools/setup/generate-test-credentials
@@ -5,8 +5,10 @@ cd "$(dirname "$0")"/../..
 
 email=iago@zulip.com
 
+mkdir -p var/casper
+
 password=$(python manage.py print_initial_password "$email" | fgrep "$email" | awk '{ print $2 }')
-cat > frontend_tests/casper_lib/test_credentials.js <<EOF
+cat > var/casper/test_credentials.js <<EOF
 // Generated by tools/setup/generate-test-credentials
 var test_credentials = {default_user: {username: '$email', password: '$password'}};
 try { exports.test_credentials = test_credentials; } catch (e) {}
